+++ CHANGELOG.md
@@ -12,6 +12,10 @@
 
 #### Builtin
 - **Breaking Change**: return value of ``Job#pid`` method
+- **Breaking Change**: some string method handle grapheme cluster
+  - ``String#count`` count grapheme clusters in string
+    ``String#charAt`` get grapheme cluster at specified position
+  - iterate grapheme cluster in string
 - propagate intrenal error as exception from regex method
   - ``left =~ right``, ``left !~ right``
   - ``Regex#match``, ``Regex#replace``